+++ NEWS.md
@@ -1,6 +1,7 @@
 
-# cdata 0.7.2 2018/06/25
+# cdata 0.7.2 2018/07/07
 
+ * switch local ops to data.table implementation.
  * re-export more of wrapr
  * move db fns to rquery.